mbref: make distribution
Use 'git archive' to producing a tarball directly from the
git master branch. If you do this however, you end up with
the entire contents of the git branch, including potentially
undesirable files like: .gitignore, doc/ or the big stuff
in reference-designs/

So we use the .gitattributes file to exclude the most
unnecessary files and directories.

Signed-off-by: Stephan Linz <linz@li-pro.net>